refactor(ai): SRP-decompose GoldenPathSynthesizer — extract the remaining 4 clusters into focused modules (#14281)
Finish the GoldenPathSynthesizer SRP decomposition: extract the remaining inline-logic
clusters into focused sibling modules, behavior-preserving via stable static delegating
shims (the public API is unchanged).

New modules (ai/services/graph/):
- agentFamilyResolution.mjs     — identity/family resolution from the roster + PR self-id + cross-family review
- computedGoldenPathRouting.mjs — the Computed Golden Path routing-decision surface + routed/empty/failure render sections
- activePrCycleSection.mjs      — Active PR Cycle State handoff rendering
- frontierConsolidation.mjs     — frontier GUIDE-edge pruning + recency-ordered summary reads + Consolidation Gaps

The issue-focus clusters were already extracted to issueFocusSections.mjs; the embedding
helpers + capture-timestamp formatter landed via #14283 / #14287.

GoldenPathSynthesizer.mjs: 1504 -> 1106 lines — now the synthesizeGoldenPath orchestration
core plus the stable-API shim layer. Pure moves + import rewiring; the routing-vs-visibility
section boundary is preserved (no behavior change).

Verified locally: GoldenPathSynthesizer.spec 36/36; QueryService.queryDocuments + MemoryService.Frontier 14/14 (graph-dir now 17 files, still within the limit:25 exact-anchor headroom).
